--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/sample/data/mappers/SampleDtoMapper.kt
@@ -1,17 +1,0 @@
-package com.whatever.caramel.feat.sample.data.mappers
-
-import com.whatever.caramel.feat.sample.data.dto.response.SampleDetailDto
-import com.whatever.caramel.feat.sample.data.dto.response.SampleGetMethodResponseDto
-import com.whatever.caramel.feat.sample.domain.SampleDetail
-import com.whatever.caramel.feat.sample.domain.SampleModel
-
-fun SampleGetMethodResponseDto.toSampleModel(): SampleModel =
-    SampleModel(
-        name = this.name,
-        localDateTime = this.localDateTime,
-        detail = this.detail.toSampleDetail(),
-        detailArray = this.detailArray.map { it.toSampleDetail() }
-    )
-
-fun SampleDetailDto.toSampleDetail(): SampleDetail =
-    SampleDetail(description = this.description)
