+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/sample/domain/SampleModel.kt
@@ -1,5 +1,8 @@
 package com.whatever.caramel.feat.sample.domain
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class SampleModel(
     val name: String = "",
     val localDateTime: String = "",
@@ -7,6 +10,7 @@ data class SampleModel(
     val detailArray: List<SampleDetail> = emptyList(),
 )
 
+@Serializable
 data class SampleDetail(
     val description: String = ""
 )
