+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/sample/domain/SampleRepository.kt
@@ -4,4 +4,8 @@ interface SampleRepository {
 
     suspend fun getSampleData(): SampleModel
 
+    suspend fun saveSampleDataToLocal(data : SampleModel)
+
+    suspend fun getSampleDataFromLocal() : List<SampleModel>
+
 }
